change ray calc on double

+++ src/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: vpoirot <vpoirot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 11:01:31 by vpoirot           #+#    #+#             */
-/*   Updated: 2024/01/15 10:54:27 by vpoirot          ###   ########.fr       */
+/*   Updated: 2024/01/15 12:51:03 by vpoirot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,7 +83,7 @@ typedef struct s_data
 	char			**textures;
 	char			**map_flat;
 	int				*player_pos;
-	double			*rays;
+	int				*rays;
 	mlx_t			*mlx_ptr;
 	struct s_img	imgs;
 	struct s_colors	colors;
@@ -118,7 +118,7 @@ void	ft_rotate_point(double *dir_x, double *dir_y, double rotspeed);
 int		pos_mouse(t_data *data);
 
 // raycast.c
-double	raygun(t_data *data, double pos_x, double pos_y,
+int		raygun(t_data *data, double pos_x, double pos_y,
 			double dir_x, double dir_y);
 void	ray_view(t_data *data);
